[fasterize/FasterizeEngine#862] Unloadable scripts should not stop execution of deferjs.
On Chrome, handlers of the dummy script are not called if the script cannot be loaded.
The error handler of the cloned script is called in this case.
The dummy script is used by the prefetch feature which has been removed
since [fasterize/FasterizeEngine#775].
So this commit removes the dummy script and adds the handlers to the
cloned script.